requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.10.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
gunicorn==20.1.0
django-heroku==0.3.1
drf-yasg